Exception handling - delete

+++ src/main/java/com/unbfjmc/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.unbfjmc.course.entities.User;
 import com.unbfjmc.course.repositories.UserRepository;
+import com.unbfjmc.course.services.exceptions.DatabaseException;
 import com.unbfjmc.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -29,9 +31,17 @@ public class UserService {
 		return repository.save(obj);
 	}
 	
-	public void delete(Long id) {
-		repository.deleteById(id);
-	}
+    public void delete(Long id) {
+        try {
+            if (repository.existsById(id)) {
+                repository.deleteById(id);			
+            } else {				
+                throw new ResourceNotFoundException(id);			
+            }		
+        } catch (DataIntegrityViolationException e) {			
+            throw new DatabaseException(e.getMessage());		
+        }	
+    } 
 	
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
